Grading calculator GUI (C++)
I created a Grade calculator for our CS102 class that would give the user their overall final grade based on the scores they receives for their quizzes, final exam, and assignments. I also created the GUI so the user would be able to manually enter in the text box the percentage of their score, and from there, it would do the calculations based off the algorithm I had implemented which used the weights from the classes grading scale. The coding language I used was c++